requirements.txt
aiopg==0.16.0
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.3
pylint==2.3.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
websockets==8.0.1
wrapt==1.11.2
